add manage listing and user action auth protection

+++ app/Http/Controllers/ListingController.php
@@ -41,6 +41,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         } //file upload
 
+        $formFields['user_id'] = auth()->id();
+
         //finally store the filled data
         Listing::create($formFields);
 
@@ -58,12 +60,22 @@ class ListingController extends Controller
 
     public function edit(Listing $listing)
     {
-        return view('listings.edit',['listing'=> $listing]);
+        return view('listings.edit', ['listing' => $listing]);
+    }
+
+    //Manage
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 
     //Store 
     public function update(Request $request, Listing $listing)
     {
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -86,11 +98,13 @@ class ListingController extends Controller
     }
 
     //Delete (Destroy)
-    public function destroy(Listing $listing){
+    public function destroy(Listing $listing)
+    {
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
-
-
 }
